requirements.txt
numpy
torch>=1.1
numpy
pyyaml
torchvision
matplotlib
tensorboardX
packaging
scipy
